Refactor PublicModel class and remove unused code
